perfprof: controller: implicit wait for cgroups v1
rather than inject an explicit wait in the reconcile loop,
we check that the MachineConfig includes the required cgroups V1
settings, exiting from the reconcile loop if not.

In addition, all the tuning features depend on cgroups v1.
Hence, we set the perfprofile status to Degraded until the node(s)
are configured to v1.

Last but not least, it must be noted that this approach depends
on implementation details pertaining to MCO. We are thus adding
an implicit dependency between the two components.

Signed-off-by: Francesco Romani <fromani@redhat.com>